backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# OpenAI (バージョン固定 + 依存関係明示)
openai==1.54.3
httpx==0.27.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.36

# Configuration
pyyaml==6.0.2
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.6.0

# Utilities
redis==5.0.1